refactor: CenterLayout 컴포넌트 type, style 파일 제거
