--- src/components/layout/CenterLayout/CenterLayout.type.ts
@@ -1,6 +1,0 @@
-import { AvailableColor } from 'styles/type';
-
-export type CenterLayoutProps = {
-  headerBackgroundColor?: AvailableColor;
-  mainBackgroundColor?: AvailableColor;
-};
